requirements.txt
GitPython
python-cjson
uncertainties
nose
sphinx
git+https://github.com/kreczko/hepshell.git
git+https://github.com/raggleton/htcondenser.git
six
